+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Change self-enum color components with Dart builtin enums
 - Remove unused enum items (RGB, RGBA, HEX)
-- Move color components from `package:fff/parser.dart` to `package:fff/color.dart`
+- Move color components from `package:flutter_color_hex/parser.dart` to `package:flutter_color_hex/color.dart`
 - Restored tests to work with the new Dart library
 - Added the ability to specify a pattern when translating a color into a string